+++ NEWS.md
@@ -1,5 +1,9 @@
 # arima2 (development version)
 
+# arima2 3.0.3
+
+* Added `aicc` option to `aicTable`, and removed superfluous arguments. 
+
 # arima2 3.0.2
 
 * Added `...` argument to `aicTable` function. 
@@ -7,7 +11,7 @@
 # arima2 3.0.1
 
 * Added arXiv paper to package description
-* Added parathesis to function names in package description. 
+* Added parenthesis to function names in package description. 
 
 # arima2 3.0.0